refactor: reorganizes symbol provider classes
